moving stuff around n shi
