+++ blockbuster/lane_handlers.go
@@ -1,7 +1,6 @@
 package blockbuster
 
 import (
-	"context"
 	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -152,48 +151,5 @@ func (l *LaneConstructor[C]) DefaultCheckOrderHandler() CheckOrderHandler {
 func DefaultMatchHandler() MatchHandler {
 	return func(ctx sdk.Context, tx sdk.Tx) bool {
 		return true
-	}
-}
-
-// DefaultTxPriority returns a default implementation of the TxPriority. It prioritizes
-// transactions by their fee.
-func DefaultTxPriority() TxPriority[string] {
-	return TxPriority[string]{
-		GetTxPriority: func(goCtx context.Context, tx sdk.Tx) string {
-			feeTx, ok := tx.(sdk.FeeTx)
-			if !ok {
-				return ""
-			}
-
-			return feeTx.GetFee().String()
-		},
-		Compare: func(a, b string) int {
-			aCoins, _ := sdk.ParseCoinsNormalized(a)
-			bCoins, _ := sdk.ParseCoinsNormalized(b)
-
-			switch {
-			case aCoins == nil && bCoins == nil:
-				return 0
-
-			case aCoins == nil:
-				return -1
-
-			case bCoins == nil:
-				return 1
-
-			default:
-				switch {
-				case aCoins.IsAllGT(bCoins):
-					return 1
-
-				case aCoins.IsAllLT(bCoins):
-					return -1
-
-				default:
-					return 0
-				}
-			}
-		},
-		MinValue: "",
 	}
 }
