+++ blockbuster/lane_mempool.go
@@ -36,6 +36,49 @@ type (
 	}
 )
 
+// DefaultTxPriority returns a default implementation of the TxPriority. It prioritizes
+// transactions by their fee.
+func DefaultTxPriority() TxPriority[string] {
+	return TxPriority[string]{
+		GetTxPriority: func(goCtx context.Context, tx sdk.Tx) string {
+			feeTx, ok := tx.(sdk.FeeTx)
+			if !ok {
+				return ""
+			}
+
+			return feeTx.GetFee().String()
+		},
+		Compare: func(a, b string) int {
+			aCoins, _ := sdk.ParseCoinsNormalized(a)
+			bCoins, _ := sdk.ParseCoinsNormalized(b)
+
+			switch {
+			case aCoins == nil && bCoins == nil:
+				return 0
+
+			case aCoins == nil:
+				return -1
+
+			case bCoins == nil:
+				return 1
+
+			default:
+				switch {
+				case aCoins.IsAllGT(bCoins):
+					return 1
+
+				case aCoins.IsAllLT(bCoins):
+					return -1
+
+				default:
+					return 0
+				}
+			}
+		},
+		MinValue: "",
+	}
+}
+
 // NewConstructorMempool returns a new ConstructorMempool.
 func NewConstructorMempool[C comparable](txPriority TxPriority[C], txEncoder sdk.TxEncoder, maxTx int) *ConstructorMempool[C] {
 	return &ConstructorMempool[C]{
